# project/parttobe/endpoints.openapi.yaml
openapi: "3.0.0"
servers:
- url: http://testserver/api
info:
  version: 1.0.0
  title: Part To
  license:
    name: MIT
paths:
...
  /api/run/start:
    post:
      summary: start a set of definitions
      operationId: runstart:post
      tags:
      - run
      requestBody:
        description: List of unique part to titles to run
        required: true
        content:
          '*':
            schema:
              type: object
              properties:
                runState:
                  type: string
                  format: RunStateId
      responses:
        '404':
          description: a frontend bug
          content:
            '*':
              schema:
                $ref: "#/components/schemas/Four04Reply"
        '500':
          description: a backend bug
          content:
            '*':
              schema:
                type: string
        '400':
          description: a frontend bug
          content:
            '*':
              schema:
                type: array
                items:
                  type: string
        '200':
          description: Expected response to a valid request
          content:
            '*':
              schema:
                $ref: "#/components/schemas/RunOperationReply"
...
  /api/run/stage:
    post:
      summary: stage a set of definitions
      operationId: runstage:post
      tags:
      - run
      requestBody:
        description: List of unique part to titles to run
        required: true
        content:
          '*':
            schema:
                $ref: "#/components/schemas/RunOperation"
      responses:
        '404':
          description: a frontend bug
          content:
            '*':
              schema:
                $ref: "#/components/schemas/Four04Reply"
        '500':
          description: a backend bug
          content:
            '*':
              schema:
                type: string
        '400':
          description: a frontend bug
          content:
            '*':
              schema:
                type: array
                items:
                  type: string
        '200':
          description: Expected response to a valid request
          content:
            '*':
              schema:
                $ref: "#/components/schemas/RunOperationReply"
components:
  schemas:
    Four04Reply:
      type: object
      properties:
        messages:
          type: array
          items:
            type: string
    RunOperationReply:
      type: object
      required:
      - runState
      properties:
        runState:
          type: string
          format: RunStateId
    RunOperation:
      type: object
      required:
        - definitions
      properties:
        runState:
          type: string
          format: RunStateId
        definitions:
          type: array
          items:
            type: string
            format: TaskDefinitionId
          minItems: 1
...